Fix select parsing # plural interpolations
With this commit select interpolations will now correctly parse `#` as
plaintext, except nested inside a plural interpolation where they'll
correctly take on that outer argument.

Technically, previously there were two separate usages of
`withPluralCtx` and when the first failed there would be no cleanup,
resulting in a dangling entry in `pluralArgNameStack`.

It's possible as of this commit that issues with `withPluralCtx` could
still bubble up and cause bugs depending upon how `interp` is used.

The fundamental issue with `withPluralCtx` is that if `p` fails then its
cleanup function, `popPluralCtx`, is never called. Observe the following
in a REPL:

> Nothing <* traceShowId (Just 2)
> Just 1 <* traceShowId (Just 2)

The applicative operation shortcircuits. This can seem counterintuitive given
we're intending to run the applicative on the right but retain the value
of that on the left. In actuality we run and retain the value on the
left first before, if successful, running and discarding the value on
the right.