Fix escaping to ICU 4.8+ spec
See the link in the comment.

Previously we parsed two single quotes anywhere in a message as an
escape sequence, however now as per spec we only do so if the opening
quote precedes a syntax character.

As a byproduct this doesn't fix the cross-message bug, but it does in
practice almost entirely eliminate it.